myproject/requirements.txt
django==4.0.6
django-ninja==0.19.1
django-redis==5.2.0
